added feature to log status of execution

+++ src/main/scala/org/wf/StatusUpdate.scala
@@ -1,6 +1,6 @@
 package org.wf
 
-case class StatusMessage(flowName: String, jobName: Option[String]=None, jobType: Option[String]=None, msg: String, exception: Option[String] = None,time:java.util.Date, tableName:Option[String] = None,tableCount:Option[Int] = None)
+case class StatusMessage(flowName: String, jobName: Option[String]=None, jobType: Option[String]=None, msg: String, exception: Option[String] = None,time:java.util.Date, tableName:Option[String] = None,tableCount:Option[Long] = None)
 
 object StatusUpdate {
